+++ src/nn/mod.rs
@@ -11,7 +11,7 @@ pub trait Layer< IN > {
 
     const NUM_PARAMETERS: usize;
 
-    fn load(&mut self, parameter: &[f64]);
+    fn load(&mut self, parameter: &[f32]);
     fn out(&self) -> &Self::OUT;
     fn forward(&mut self, input: &IN);
 }
